requirements.txt
apache-airflow-providers-amazon
s3fs
apache-airflow
pyspark==3.5.0
pytest